# docs/swagger.yaml
basePath: /
...
host: mii-unsecure.ariankordi.net
info:
  contact: {}
  description: 'This API contains endpoints to fetch Mii data from NNID, CMOC, or
    Miitomo and an endpoint to render a Mii icon. As of writing, the repo for the
    renderer server is located here: https://github.com/ariankordi/FFL-Testing/tree/renderer-server-prototype
    and for the frontend/APIs itself is located here: https://github.com/ariankordi/nwf-mii-cemu-toy/tree/ffl-renderer-proto-integrate
    If you are lurking and found this, my email address is ariankordi@ariankordi.net'
  title: Mii Renderer (REAL) Frontend and API
  version: "1.0"
paths:
...
  /miitomo_get_player_data/{player_id}:
    get:
      description: 'Get data (e.g. nn::mii::CoreData, Mii image...) from a player
        on Miitomo (Kaerutomo) by their player ID. As of writing, this is requesting
        the following API endpoint: ktp.kaeru.world/v1/player/get_player_data'
      parameters:
      - description: Player ID. This is in your friend request invite link as the
          first 16 character hex string after friend_code/
        format: string, 16-character hex
        in: path
        name: player_id
        type: string
      - description: Target player ID. Accepts multiple comma separated values. This
          works in substitute of the player ID path parameter.
        format: string, 16-character hex
        in: query
        name: target_player_id
        type: string
      - description: 'Namespace. Accepts multiple comma separated values. Known values:
          own_mii, stock_mii, mii_face_image, account_info, stock_item, reported_content,
          corinth_status, star_account_info, _star_account_info, message_info'
        in: query
        name: namespace
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: Player data returned from the upstream.
          schema:
            additionalProperties: true
            type: object
        "400":
          description: Bad Request
          schema:
            type: string
        "404":
          description: Player data not found
          schema:
            type: string
        "500":
          description: Internal Server Error. Error with marshaling, obfuscating,
            or sending request.
          schema:
            type: string
        "502":
          description: Bad Gateway
          schema:
            type: string
      summary: Lookup Miitomo Player Data
      tags:
      - Fetch Mii Data
swagger: "2.0"